+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### 9.23.0
 
+* `[log]` Added wrapper for compatibility with stdlib logger
 * `[log]` Fixed race condition issue
 
 ### 9.22.3